requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.3.4
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.141.0
tenacity==8.5.0
httpx==0.27.0
python-dateutil==2.9.0.post0
nanoid==2.0.0
SQLAlchemy==2.0.30
alembic==1.13.2
tzlocal==5.2
dateparser==1.2.0
loguru==0.7.2
